added encryption module
