+++ lib/slack/smart-bot/utils.rb
@@ -32,8 +32,8 @@ require_relative 'utils/update_teams'
 require_relative 'utils/update_vacations'
 require_relative 'utils/check_vacations'
 require_relative 'utils/display_calendar'
-require_relative 'utils/encryption_get_key_iv'
-require_relative 'utils/encrypt'
-require_relative 'utils/decrypt'
+require_relative 'utils/encryption/encryption_get_key_iv'
+require_relative 'utils/encryption/encrypt'
+require_relative 'utils/encryption/decrypt'
 require_relative 'utils/get_team_members'
