+++ lib/slack/smart-bot/utils/encryption/encryption_get_key_iv.rb
@@ -1,5 +1,7 @@
 class SlackSmartBot
-    def encryption_get_key_iv
+  module Utils
+    module Encryption
+      def self.encryption_get_key_iv(config)
         if defined?(@encryption_key_built)
           key = @encryption_key_built
           iv = @encryption_iv_built
@@ -11,19 +13,21 @@ class SlackSmartBot
             key = (Socket.gethostname + config.token.reverse)[0..49]
             iv = config.token[0..15]
           end
-    
+
           #Convert from hex to raw bytes:
-          key = [key].pack('H*')
+          key = [key].pack("H*")
           #Pad with zero bytes to correct length:
           key << ("\x00" * (32 - key.length))
-        
+
           #Convert from hex to raw bytes:
-          iv = [iv].pack('H*')
+          iv = [iv].pack("H*")
           #Pad with zero bytes to correct length:
-          iv << ("\x00" * (16 - iv.length))  
+          iv << ("\x00" * (16 - iv.length))
           @encryption_key_built = key
           @encryption_iv_built = iv
         end
         return key, iv
-    end      
+      end
+    end
+  end
 end
