+++ lib/slack/smart-bot/utils/get_personal_settings.rb
@@ -5,7 +5,7 @@ class SlackSmartBot
     files = Dir.glob(File.join(config.path, "personal_settings", "ps_*.yaml"))
     files.each do |file|
       if !defined?(@datetime_personal_settings_file) or !@datetime_personal_settings_file.key?(file) or @datetime_personal_settings_file[file] != File.mtime(file)
-        user_personal_settings = YAML.load(decrypt(File.read(file)))
+        user_personal_settings = YAML.load(Utils::Encryption.decrypt(File.read(file),config))
         @personal_settings[File.basename(file).gsub("ps_","").gsub(".yaml","")] = user_personal_settings
         @datetime_personal_settings_file[file] = File.mtime(file)
       end
