+++ lib/slack/smart-bot/utils/update_teams.rb
@@ -13,7 +13,7 @@ class SlackSmartBot
       team_file = File.join(config.path, "teams", "t_#{team}.yaml")
       File.open(team_file, 'w') {|file|
         file.flock(File::LOCK_EX)
-        file.write(encrypt(@teams[team].to_yaml))
+        file.write(Utils::Encryption.encrypt(@teams[team].to_yaml, config))
         file.flock(File::LOCK_UN)
       }
       @datetime_teams_file[team_file] = File.mtime(team_file)
