+++ lib/slack/smart-bot/utils/update_vacations.rb
@@ -10,7 +10,7 @@ class SlackSmartBot
 
     File.open(vacations_file, 'w') {|file|
       file.flock(File::LOCK_EX)
-      file.write(encrypt(@vacations[user.name].to_yaml))
+      file.write(Utils::Encryption.encrypt(@vacations[user.name].to_yaml, config))
       file.flock(File::LOCK_UN)
     }
     @datetime_vacations_file[vacations_file] = File.mtime(vacations_file)
